[FLINK-6134] Set UUID(0L, 0L) as default leader session id
Before the default leader session id was null in the standalone case. However, in the ZooKeeper
case null indicated that there was no active leader available. With this commit, the default
leader id will be set to UUID(0L, 0L). This allows the uniform treatment of null denoting that
there is no active leader across the standalone and the ZooKeeper case.

With this change, the FlinkActors will now ignore all LeaderSessionMessages if the actors's
leader id field is null. This indicates that the FlinkActor does not know the current leader.

This closes #3578.