[hotfix] Remove validateRunsInMainThread from TaskExecutor to fix TaskExecutorTest
Currently, the TestingSerialRpcService does not play well together with the
MainThreadValidatorUtil which assumes that rpc calls are dispatched to a mailbox
from where they are picked up. In order to support the TestingSerialRpcService
we will have to extend the MainThreadValidatorUtil to allow entering the main thread
if the RpcEndpoint already runs in the context of the main thread.